% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_extension.R
\name{correct_extension}
\alias{correct_extension}
\title{Corrects file extension}
\usage{
correct_extension(file, mime = get_content_type(file), ...)
}
\arguments{
\item{file}{A file to read in}

\item{mime}{mime type}

\item{...}{not used}
}
\value{
Correct file name
}
\description{
This function uses get_content_type() to append the correct file extension to the file name
the user wants to read in. This is useful in cases where the file extension is wrong, or where
the file extension was removed by mistake.
}
\examples{
\dontrun{
file <- system.file("dataset","wrong_is_xlsx.csv",
package = "cleanser")
correct_extension(file)
}
}
