% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_encoding.R
\name{detect_encoding}
\alias{detect_encoding}
\title{guess encoding (brute force)}
\usage{
detect_encoding(base, dec = ",", skip = 0, header = TRUE,
  sep = "\\t", nrows = 10, info = FALSE)
}
\arguments{
\item{base}{path to file}

\item{dec}{decimal separator character}

\item{skip}{number of rows to skip}

\item{header}{a logical value indicating whether the file contains the names of the variables as its first line.}

\item{sep}{the field separator character.}

\item{nrows}{number of rows to import}

\item{info}{booleen add more information}
}
\description{
guess encoding (brute force)
}
\examples{
library(dplyr)
file <- system.file("dataset","boys.csv",
                    package = "cleanser")
detect_encoding(file,sep = ",",nrows = 100)\%>\% arrange(desc(row)) \%>\% head()



}
