% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_encoding.R
\name{detect_encoding_brute}
\alias{detect_encoding_brute}
\title{guess encoding (brute force) with all dec and sep parameters}
\usage{
detect_encoding_brute(base, nrows = 10)
}
\arguments{
\item{base}{path to file}

\item{nrows}{number of row to import}
}
\description{
guess encoding (brute force) with all dec and sep parameters
}
\examples{
\dontrun{
library(dplyr)                   
file <- system.file("dataset","boys.csv",
                    package = "cleanser")
detect_encoding_brute(file) \%>\% arrange(desc(row)) \%>\% head()
}
}
