% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/force_read.R
\name{force_read}
\alias{force_read}
\title{Reads text files in R}
\usage{
force_read(file, ...)
}
\arguments{
\item{file}{A file to read in.}

\item{...}{Additional argument to pass down to methods.}
}
\value{
A data frame.
}
\description{
This function is a wrapper around read_delim and read.table.
}
