% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_available_encoding.R
\name{get_available_encoding}
\alias{get_available_encoding}
\title{Return all avaible encoding option}
\usage{
get_available_encoding()
}
\description{
Return all avaible encoding option
}
\examples{
get_available_encoding()
}
