% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ni.R
\name{\%ni\%}
\alias{\%ni\%}
\title{not in}
\usage{
x \%ni\% table
}
\arguments{
\item{x}{a value to}

\item{table}{a vector to parse}
}
\value{
a booleen
}
\description{
not in
}
\examples{
"a" \%ni\% letters
"coucou" \%ni\% letters

}
