% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_separator.R
\name{guess_separator}
\alias{guess_separator}
\title{guess txt file separator}
\usage{
guess_separator(file, separator = c(",", ";", " ", "\\t"),
  n_max = 1000)
}
\arguments{
\item{file}{path to csv or txt file}

\item{separator}{vector of possible separator to test}

\item{n_max}{number of row to parse}
}
\value{
sorted vector of possible separator
}
\description{
guess txt file separator
}
\examples{
file <- system.file("dataset","demo.csv",package = "cleanser")
guess_separator(file)
file2 <- system.file("dataset","demo2.csv",package = "cleanser")
guess_separator(file2)
}
