% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type_mime.R
\name{is_csv}
\alias{is_csv}
\title{Checks if a file is a csv file}
\usage{
is_csv(type_mime)
}
\arguments{
\item{type_mime}{character The type mime of the file to check}
}
\description{
Checks if a file is a csv file
}
