% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type_mime.R
\name{is_excel}
\alias{is_excel}
\title{Checks if a file is an excel file}
\usage{
is_excel(type_mime)
}
\arguments{
\item{type_mime}{character The type mime of the file to check}
}
\description{
Checks if a file is an excel file
}
