% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_likely_date.R
\name{is_likely_date}
\alias{is_likely_date}
\title{Signal to the user that a variable is likely a date}
\usage{
is_likely_date(dataframe, tol = 5)
}
\arguments{
\item{dataframe}{A dataframe.}

\item{tol}{The maximum probability tolerated for this column to be a date. The lower, the stricter.}
}
\value{
A boolean vector.
}
\description{
This function takes a dataframe as input and returns a boolean vector of positions where
it is likely that a column is a date.
}
\examples{
\dontrun{
library(tidyverse)
data(storms)
storms <- storms \%>\% 
  mutate(date = paste(year, month, day, sep = "/"))

is_likely_date(storms)
}
}
