% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_likely_date.R
\name{is_likely_date_col}
\alias{is_likely_date_col}
\title{Is a data frame column likely a date?}
\usage{
is_likely_date_col(col, tol = 10)
}
\arguments{
\item{col}{A data frame column.}

\item{tol}{The maximum probability tolerated for this column to be a date. The lower, the stricter.}
}
\value{
TRUE or FALSE.
}
\description{
This function determines, with a certain tolerance, whether a column is likely a date.
}
\examples{
\dontrun{
library(tidyverse)
data(storms)
storms <- storms \%>\% 
  mutate(date = paste(year, month, day, sep = "/"))

is_likely_date_col(storms$date)
}
}
