% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_likely_factor.R
\name{is_likely_factor}
\alias{is_likely_factor}
\title{Signal to the user that a variable is likely a factor}
\usage{
is_likely_factor(dataframe, tol = 5)
}
\arguments{
\item{dataframe}{A dataframe}

\item{tol}{The maximum tolerance for this column to be a date. The lower, the stricter.}
}
\value{
A boolean vector.
}
\description{
This function takes a dataframe as input and returns a boolean vector of positions where
it is likely that a column is a factor. For example, a column of only 0's and 1's is likely a factor.
}
\examples{
\dontrun{
is_likely_factor(dataframe)
}
}
