% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_likely_numeric.R
\name{is_likely_numeric}
\alias{is_likely_numeric}
\title{Signal to the user that a variable is likely a numeric}
\usage{
is_likely_numeric(dataframe)
}
\arguments{
\item{dataframe}{A dataframe}
}
\value{
A boolean vector.
}
\description{
This function takes a dataframe as input and returns a boolean vector of positions where
it is likely that a column is numeric with decimal. For example, a column of only numbers with decimal is likely a numeric.
}
\examples{
\dontrun{
is_likely_numeric(dataframe)
}
}
