% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type_mime.R
\name{is_ods}
\alias{is_ods}
\title{Checks if a file is an ods file}
\usage{
is_ods(type_mime)
}
\arguments{
\item{type_mime}{character The type mime of the file to check}
}
\description{
Checks if a file is an ods file
}
