% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_type_mime.R
\name{is_rds}
\alias{is_rds}
\title{Checks if a file is a rds file}
\usage{
is_rds(type_mime)
}
\arguments{
\item{type_mime}{character The type mime of the file to check}
}
\description{
Checks if a file is a rds file
}
