% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likely_types.R
\name{likely_types}
\alias{likely_types}
\title{Signal to the user that variables are likely of the wrong type}
\usage{
likely_types(dataframe, tol = 5)
}
\arguments{
\item{dataframe}{A dataframe}

\item{tol}{Maximum number of unique values in a factor}
}
\value{
A vector of html colors, one color per type
}
\description{
This function takes a dataframe as input and returns a character vector of html
colors.
}
\examples{
\dontrun{
storms = storms \%>\%
  mutate(date = paste(year, month, day, sep = "/"))
is_likely_date(storms)
}
}
