% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_all_fileInput.R
\name{mod_all_fileInput}
\alias{mod_all_fileInput}
\alias{mod_all_file}
\title{mod_all_fileInput and mod_all_file}
\usage{
mod_all_fileInput(id, label = "Upload your file")

mod_all_file(input, output, session, stringsAsFactors = TRUE, r)
}
\arguments{
\item{id}{shiny id}

\item{label}{file input label}

\item{input}{For internal purposes}

\item{output}{For internal purposes}

\item{session}{The current session. Used internally.}

\item{stringsAsFactors}{Should strings be imported as factors? True by default.}
}
\description{
A shiny Module that imports everything

Module that imports, csv, Excel and rds files.
}
\examples{
if (interactive()) {
library(shiny)
library(cleanser)
library(shinyjs)
library(DT)
ui <- function() {
  fluidPage(
    titlePanel("cleanser"),
    useShinyjs(), 
    sidebarLayout(
      sidebarPanel(
        mod_all_fileInput("fichier")
      ),
      mainPanel(
        DTOutput("tableau")
      )
    )
  )
}


server <- function(input, output,session) {
  
  data <- callModule(mod_all_file,"fichier")
  
  output$tableau <- renderDT({data()})
}
shinyApp(ui = ui, server = server)
}
}
