% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_export.R
\name{mod_exportUI}
\alias{mod_exportUI}
\alias{mod_export}
\title{Data export module}
\usage{
mod_exportUI(id, label = "Upload your file")

mod_export(input, output, session, r)
}
\arguments{
\item{id}{shiny id}

\item{label}{file input label}

\item{input}{For internal purposes}

\item{output}{For internal purposes}

\item{session}{The current session. Used internally.}

\item{data}{The data to export}
}
\description{
A shiny Module that exports data to several formats
}
\examples{
if (interactive()) {
library(shiny)
library(cleanser)
library(shinyjs)
library(DT)
ui <- function() {
  fluidPage(
    titlePanel("Export module"),
    useShinyjs(), 
    sidebarLayout(
      sidebarPanel(
        mod_exportUI("export")
      ),
      mainPanel(
        DTOutput("tableau")
      )
    )
  )
}


server <- function(input, output,session) {
    
    reactive_iris <- reactive({
      iris
    })
    
  callModule(mod_export, "export", reactive_iris)
  
  output$tableau <- renderDT({
  iris
  })
}
shinyApp(ui = ui, server = server)
}
}
