% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_factorUI.R
\name{mod_factorUI}
\alias{mod_factorUI}
\alias{mod_factor}
\title{mod_factor and mod_factorUI  function}
\usage{
mod_factorUI(id)

mod_factor(input, output, session, r)
}
\arguments{
\item{id}{The id of the current element you are entering}

\item{input}{internal}

\item{output}{internal}

\item{session}{internal}

\item{data}{dataset as reactive}
}
\description{
Shiny Module that allows to recode,reorder, factor levels.
}
\examples{

if (interactive()){

library(shiny)
library(cleanser)
library(shinyjs)
library(DT)
library(readr)
ui <- function() {
  fluidPage(
    titlePanel("cleanser"),
    useShinyjs(),
    sidebarLayout(
      sidebarPanel(
        selectInput("go","change data",
        c("demo7_rds","boys",
         "vins",
         "vins_missing",
         "one_factor",
    "right_csv",
    "right_xls",
    "demo2_csv",
    "demo3_xlsx",
    "demo4_rds"
        ))
      ),
      mainPanel(
        mod_factorUI("demo")
      )
    )
  )
}


server <- function(input, output,session) {
  
  
  donnee <- reactive({
    
    md <- function(x){system.file("dataset", x, package = "cleanser")}
    switch (input$go,
  "boys" = readr::read_csv(md("boys.csv")),
  "vins" = readr::read_csv(md("vins.csv")),
  "vins_missing" = readr::read_csv(md("vins_missing.csv"),
   locale = locale(encoding = 'ISO-8859-1')),
  "one_factor" = readr::read_csv(md("one_factor.csv")),
  "right_csv" = readr::read_csv(md("right.csv")),
  "right_xls" = readxl::read_excel(md("right.xls"),1),
  "demo2_csv" = readr::read_delim(md("demo2.csv"),
   locale = locale(encoding = 'ASCII'), delim=";"),
  "demo4_rds" = readr::read_rds(md("demo4.rds")),
  "demo7_rds" = readr::read_rds(md("demo7.rds")),
  "demo3_xlsx" = readxl::read_excel(md("demo3.xlsx"),1)
    )
  })
  
  callModule(mod_factor,"demo",donnee)
  
}
shinyApp(ui = ui, server = server)
}

}
