% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeInput.R
\name{mod_impute}
\alias{mod_impute}
\title{mod_impute Server function}
\usage{
mod_impute(input, output, session, r)
}
\arguments{
\item{input}{For internal purposes}

\item{output}{For internal purposes}

\item{session}{The current session.}

\item{data}{A reactive expression returning the dataset to work on}
}
\description{
mod_impute Server function
}
