% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_imputeInput.R
\name{mod_imputeInput}
\alias{mod_imputeInput}
\title{mod_imputeInput UI function}
\usage{
mod_imputeInput(id)
}
\arguments{
\item{id}{The id of the current element you are entering}
}
\description{
mod_imputeInput UI function
}
\examples{
if (interactive()) {
library(cleanser)
ui <- shinyUI(
  fluidPage(
    titlePanel("Imputation module"),
    mod_imputeSpinInput("imp")

  ))

server <- function(input, output,session) {
  data <- reactive({
    boys
  })
  callModule(mod_imputeSpin,"imp", reactive(data()))
}

shiny::shinyApp(ui = ui, server = server)
}

}
