% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validateInput.R
\name{mod_validate}
\alias{mod_validate}
\title{mod_validate server function}
\usage{
mod_validate(input, output, session, r)
}
\arguments{
\item{input}{For internal purposes}

\item{output}{For internal purposes}

\item{session}{The current session.}

\item{data}{A reactive expression returning the dataset to work on}
}
\description{
mod_validate server function
}
