% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_validateInput.R
\name{mod_validateInput}
\alias{mod_validateInput}
\title{mod_validateInput UI function}
\usage{
mod_validateInput(id)
}
\arguments{
\item{id}{The id of the current element you are entering}
}
\description{
mod_validateInput UI function
}
\examples{
if (interactive()) {
library(cleanser)
ui <- shinyUI(
  fluidPage(
    titlePanel("Validation module"),
    mod_validateInput("val_mod")
    
  ))

server <- function(input, output,session) {
  data <- reactive({
    boys
  })
  callModule(mod_validate,"val_mod", reactive(data()))
}

shiny::shinyApp(ui = ui, server = server)

}
}
