% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safe_read.R
\name{safe_read_ods}
\alias{safe_read_ods}
\title{Safely read in data}
\usage{
safe_read_ods(...)
}
\arguments{
\item{...}{arguments to pass down to methods}
}
\description{
These functions read in data safely by wrapping import function around tryCatch
}
