% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mode.R
\name{sample_mode}
\alias{sample_mode}
\title{Returns the mode of a vector}
\usage{
sample_mode(x)
}
\arguments{
\item{x}{A list of values}
}
\value{
Returns the mode of x, preserving the type of x
}
\description{
Returns the mode of a vector
}
\examples{
x <- c(3,3,3,4,5)
sample_mode(x)
}
