% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_date.R
\name{to_date}
\alias{to_date}
\title{Converts a column to a date column}
\usage{
to_date(col)
}
\arguments{
\item{col}{A data frame column.}
}
\value{
a data frame column in the POSIX, POSIXct or Date format.
}
\description{
This function transforms a column into a Date if the format is recognizable.
}
\examples{
\dontrun{
library(tidyverse)
data(storms)
storms <- storms \%>\% 
  mutate(date = paste0(year, month, day))

to_date(storms$date)
}
}
