% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_numeric.R
\name{to_numeric}
\alias{to_numeric}
\title{Converts a data frame column of characters to numeric}
\usage{
to_numeric(col)
}
\arguments{
\item{col}{A column of characters to convert to numeric}
}
\value{
The column converted into a numeric
}
\description{
This function takes a string such as "32,486.64 €" and returns the numeric 32486.64
}
\examples{
x <- c("32,343,345.9 euros","12,749.548 euros","32,489.545 euros","143,564.95 euros")
to_numeric(x)
}
