% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_with_this_extension.R
\name{with_this_extension}
\alias{with_this_extension}
\title{Create a copy of a file with the correct extension}
\usage{
with_this_extension(file, correct_extension, ...)
}
\arguments{
\item{file}{a file containing the data}

\item{correct_extension}{character the correct extension}

\item{...}{further arguments to pass down to methods}
}
\value{
Nothing; this function is used for its side effect (create a temporary copy of the file
with the correct extension)
}
\description{
Create a copy of a file with the correct extension
}
